# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PRLifts API
  version: 1.0.0
  description: |
    PRLifts backend API. All endpoints require authentication unless marked public.
    Authentication uses Supabase JWT passed as Bearer token in Authorization header.
    All timestamps are UTC ISO 8601. All IDs are UUIDs.
    Async AI operations return HTTP 202 with a job_id — poll GET /v1/jobs/{id} for result.
  contact:
    name: PRLifts Engineering
servers:
  - url: https://api.prlifts.app/v1
    description: Production
  - url: https://api-staging.prlifts.app/v1
    description: Staging
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-04-25T14:30:00Z"
...
    ErrorResponse:
      type: object
      required: [error_code, message, request_id]
      properties:
        error_code:
          type: string
          description: Machine-readable error code. See ERROR_CATALOG.md for full list.
          example: "workout_not_found"
        message:
          type: string
          description: Human-readable message safe to display to users.
          example: "This workout could not be found."
        request_id:
          type: string
          description: Correlation ID for support queries. Searchable in Railway logs.
          example: "550e8400-e29b-41d4-a716-446655440000"
    WeightUnit:
      type: string
      enum: [kg, lbs]
    WorkoutStatus:
      type: string
      enum: [in_progress, paused, partial_completion, completed]
    WorkoutFormat:
      type: string
      enum: [weightlifting, cardio, mixed, other]
    WorkoutType:
      type: string
      enum: [ad_hoc, planned]
    WorkoutLocation:
      type: string
      enum: [gym, home, outdoor, other]
      nullable: true
    SetType:
      type: string
      enum: [normal, warmup, dropset, failure, pr]
    WeightModifier:
      type: string
      enum: [none, assisted, weighted]
...
    MuscleGroup:
      type: string
      enum:
        - upper_chest
        - mid_chest
        - lower_chest
        - upper_back
        - lower_back
        - shoulders
        - biceps
        - triceps
        - quads
        - hamstrings
        - calves
        - glutes
        - abs
        - obliques
        - full_body
    ExerciseCategory:
      type: string
      enum: [strength, cardio, flexibility, bodyweight, mobility, saq, rehab]
    ExerciseEquipment:
      type: string
      enum: [barbell, dumbbell, kettlebell, machine, cable, bodyweight, cardio_machine, other]
...
    Exercise:
      type: object
      required: [id, name, category, muscle_group, equipment, is_custom, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          maxLength: 200
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        muscle_group:
          $ref: '#/components/schemas/MuscleGroup'
        secondary_muscle_groups:
          type: array
          items:
            $ref: '#/components/schemas/MuscleGroup'
          nullable: true
        equipment:
          $ref: '#/components/schemas/ExerciseEquipment'
        instructions:
          type: string
          nullable: true
        demo_url:
          type: string
          format: uri
          nullable: true
          description: ExerciseDB GIF or video URL. Null for custom exercises.
        is_custom:
          type: boolean
        created_by:
          $ref: '#/components/schemas/UUID'
          nullable: true
          description: Null for ExerciseDB exercises. User ID for custom exercises.
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    Workout:
      type: object
      required: [id, user_id, status, type, format, started_at, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        user_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          nullable: true
          maxLength: 200
        notes:
          type: string
          nullable: true
          maxLength: 5000
        status:
          $ref: '#/components/schemas/WorkoutStatus'
        type:
          $ref: '#/components/schemas/WorkoutType'
        format:
          $ref: '#/components/schemas/WorkoutFormat'
        started_at:
          $ref: '#/components/schemas/Timestamp'
        completed_at:
          $ref: '#/components/schemas/Timestamp'
          nullable: true
        duration_seconds:
          type: integer
          nullable: true
          description: Stored explicitly for query performance. Derived from completed_at - started_at.
        location:
          $ref: '#/components/schemas/WorkoutLocation'
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    CreateWorkoutRequest:
      type: object
      required: [type, format]
      properties:
        name:
          type: string
          nullable: true
          maxLength: 200
        type:
          $ref: '#/components/schemas/WorkoutType'
        format:
          $ref: '#/components/schemas/WorkoutFormat'
        location:
          $ref: '#/components/schemas/WorkoutLocation'
        client_started_at:
          $ref: '#/components/schemas/Timestamp'
          description: Client-reported start time. Server assigns server_received_at for conflict resolution.
    UpdateWorkoutRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/WorkoutStatus'
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        completed_at:
          $ref: '#/components/schemas/Timestamp'
          nullable: true
    WorkoutExercise:
      type: object
      required: [id, workout_id, exercise_id, order_index, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        workout_id:
          $ref: '#/components/schemas/UUID'
        exercise_id:
          $ref: '#/components/schemas/UUID'
        exercise:
          $ref: '#/components/schemas/Exercise'
          description: Eagerly loaded for display convenience.
        order_index:
          type: integer
          minimum: 0
        notes:
          type: string
          nullable: true
          maxLength: 2000
        rest_seconds:
          type: integer
          nullable: true
          minimum: 0
          maximum: 3600
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    WorkoutSet:
      type: object
      required: [id, workout_exercise_id, set_number, set_type, weight_modifier, is_completed, created_at, updated_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        workout_exercise_id:
          $ref: '#/components/schemas/UUID'
        set_number:
          type: integer
          minimum: 1
          maximum: 100
        set_type:
          $ref: '#/components/schemas/SetType'
        weight:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 2000
          description: Null for pure bodyweight exercises.
        weight_unit:
          $ref: '#/components/schemas/WeightUnit'
          nullable: true
        weight_modifier:
          $ref: '#/components/schemas/WeightModifier'
        modifier_value:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 500
        modifier_unit:
          $ref: '#/components/schemas/WeightUnit'
          nullable: true
        reps:
          type: integer
          nullable: true
          minimum: 0
          maximum: 1000
        duration_seconds:
          type: integer
          nullable: true
          minimum: 0
          maximum: 86400
        distance_meters:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 100000
        calories:
          type: integer
          nullable: true
          minimum: 0
          maximum: 10000
        rpe:
          type: integer
          nullable: true
          minimum: 1
          maximum: 10
        is_completed:
          type: boolean
        notes:
          type: string
          nullable: true
          maxLength: 2000
        is_personal_record:
          type: boolean
          description: True if this set triggered a PersonalRecord. Populated after completion.
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    CursorPaginatedWorkoutResponse:
      type: object
      description: >
        Cursor-based paginated response for GET /v1/workouts.
        See ARCHITECTURE.md Decision 94. Pass next_cursor as the cursor
        query parameter to retrieve the next page.
      required: [data, has_more]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Workout'
        next_cursor:
          type: string
          nullable: true
          description: >
            Opaque base64-encoded cursor for the next page.
            Null when has_more is false.
        has_more:
          type: boolean
          description: True when additional records exist beyond this page.
...
security:
  - BearerAuth: []
paths:
...
  /workouts:
    get:
      tags: [workouts]
      summary: List workouts for current user
      operationId: listWorkouts
      description: >
        Returns workouts ordered by created_at DESC, id DESC.
        Uses cursor-based (keyset) pagination — see ARCHITECTURE.md Decision 94.
        Pass next_cursor from the previous response as the cursor parameter to
        retrieve the next page. Filters are applied before cursor positioning;
        pass the same filter values on every page of a paginated sequence.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Maximum number of workouts to return. Must be 1–100.
        - name: cursor
          in: query
          schema:
            type: string
          description: >
            Opaque pagination cursor from a previous response's next_cursor field.
            Omit to start from the first page.
        - name: format
          in: query
          schema:
            $ref: '#/components/schemas/WorkoutFormat'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkoutStatus'
      responses:
        '200':
          description: Cursor-paginated workout list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPaginatedWorkoutResponse'
        '400':
          description: Invalid cursor or limit out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [workouts]
      summary: Create a new workout
      operationId: createWorkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkoutRequest'
      responses:
        '201':
          description: Created workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
  /workouts/{workout_id}:
    parameters:
      - name: workout_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
    get:
      tags: [workouts]
      summary: Get workout by ID
      operationId: getWorkout
      responses:
        '200':
          description: Workout with exercises and sets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Workout'
                  - properties:
                      exercises:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/WorkoutExercise'
                            - properties:
                                sets:
                                  type: array
                                  items:
                                    $ref: '#/components/schemas/WorkoutSet'
        '403':
          description: Not owned by current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [workouts]
      summary: Update workout
      operationId: updateWorkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkoutRequest'
      responses:
        '200':
          description: Updated workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'

  # ── Workout Exercises ─────────────────────────────────────────────────────
...